Allow colors in arc commands on windows
Summary:
This dosen't work on cmd since this must have been disabled for a reason
but git for windows is basicaly like git on linux so there is no reason
for color to be disabled on bash type programes.

This allows colors in arc so it is not blank, makes it colorful.

Change-Id: Icadeace0f57ac57b2b9067932358d0d98385c4a4

Reviewers: greg, demon, Luke081515, mmodell

Reviewed By: Luke081515, mmodell

Differential Revision: https://phabricator.wikimedia.org/D222